requirements.txt
requests
PyYAML
websockets
pycryptodome
numpy
pandas
yfinance